Add 2 player support, directions and cleanup
There are now 2 actors that can be controlled independently.
Argueably one of the most important features of a fighting game
is someone to fight, so this is cause for celebration! Woooooo

With the 2 characters comes a need for them to properly look
the right way. They now face eachother all the time. To test
this I disabled to the collison of the characters so they can
pass through eachother(done through the inspector). This system
isn't perfect, there is most likely a better way, but hey, it
works, and it should still work in the air. This is all done by
keeping track of the distance between the players, when that
distance goes from positive to negative, or the opposite, it flips
the sprite and direction.

Cleanup... I might've cleaned up too much in the code tbh. I
just went at it with an axe, comments, random unused print
statements, more comments, unused functions, variables not needed
anymore, and some other stuff I can't remember anymore. I did try
and keep the entire structure of the program the same. But I also
did change a good amount from a structural standpoint. I rearranged
the player grouping of nodes. I really don't know if it fully makes
sense from a Godot way of doing things, but it works. Also renamed
some stuff like the main scene is now refered to as such instead of
node_2d.tscn.

In the sense of cleanup and all that, I worked with the movement a
good amount. I don't know why but the state tutorial from Godot
showed a very weird understanding of vectors that just felt wrong.
It also didn't include anything about direction because it wasn't
important to the scope of that tutorial. It's still not perfect.
The movespeed for forwards and backwards can now be set from the
inspector too!

The bad news. I did kind of break going into theidle state. It's
detailed in the comments. The way of detecting the movespeed made
the movement really weird or just flat out break, at least on my
devices, so I removed it, since for now the idle state does nothing.