input_handling.gd now reads from persistent_state.gd's buffer
Hooray! I got an input buffer to work, and to read inputs consistently and cleanly

Its already got issues :(

First off, the way it works is that now instead of checking every individual input for left or right directions, we now just hold the player's current direction, and assume they are facing right. If they aren't, we flip them so they are.

The issues that we run into now is that Player2 moves backwards on P1's forward and back movement, and doesn't read from their own movement keys. This will be next on the todo list and be merged back with main as soon as it is fixed.