added a camera node
I added a camera node which position will always be in between of the players, (pos1+pos2)/2, (After some tests I changed the code for the position.y of the camera to (posy1+posy2)/3 so that the camera stays a bit closer to the ground). There is also some limited zoom in and out and some offset.y based on the zoom, so that nothing bellow the ground is shown. A sprite2D was added to the ground as a reference for the position.